Moved map loading responsibilities from Map
MapLoader now contains the sole responsibility for loading a map from
an XML object. The MapLoader must be used to create a Map object by any
client.
