+++ routeFinder/control/SearchRunner.java
@@ -6,6 +6,7 @@ import processing.core.PApplet;
 import processing.data.XML;
 import routeFinder.model.AStarSearch;
 import routeFinder.model.Map;
+import routeFinder.model.MapLoader;
 import routeFinder.model.Point;
 
 public class SearchRunner {
@@ -73,7 +74,8 @@ public class SearchRunner {
 	}
 	
 	public void openMap(XML mapData, int width, int height) {
-		map = new Map(mapData, width, height);
+		MapLoader loader = new MapLoader(width, height);
+		map = loader.loadMap(mapData);
 	}
 	
 	/*
