+++ routeFinder/model/.gitignore
@@ -6,3 +6,4 @@
 /AStarSearch.class
 /SearchPoint.class
 /MapLoader.class
+/Bounds.class
